fix: keep quota attribution below smell limits
